+++ test/org/soulspace/qclojure/domain/math_test.clj
@@ -238,6 +238,5 @@
 
 (comment
   (run-tests)
-
   ;
   )